fix(server.js env): added ADMIN_JWT_TOKEN environment and update README with instructions for create
Hi!!! Thank you for reporting this issue. The problem was caused by my inattention when migrating
from version 3.0.x to version 3.1 of strapi and not making the necessary changes. The bug related to
this problem has been fixed according to the strapi guide and the README has been updated with
instructions to create the token, check it. I hope this helps.
fix #84